backend/requirements.txt
# Core Framework (Compatible versions for Python 3.12)
langchain>=0.3.10,<1.0.0
langchain-core>=0.3.15,<0.4.0
langchain-openai>=0.2.9,<0.4.0
langgraph>=0.2.45,<1.0.0

# API Framework
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
pydantic>=2.9.0
pydantic-settings>=2.6.0

# OpenAI SDK for Sora API
openai>=1.54.0,<2.0.0

# Utilities
python-dotenv==1.0.1
httpx==0.27.2
aiofiles==24.1.0

# Database (SQLite with SQLAlchemy)
sqlalchemy==2.0.35
alembic==1.13.3
